Move pages to test support package
