+++ collect_app/src/androidTest/java/org/odk/collect/android/regression/FillBlankFormTest.java
@@ -13,12 +13,12 @@ import org.junit.rules.RuleChain;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.R;
 import org.odk.collect.android.activities.FormEntryActivity;
-import org.odk.collect.android.espressoutils.pages.BlankFormSearchPage;
-import org.odk.collect.android.espressoutils.pages.ExitFormDialog;
-import org.odk.collect.android.espressoutils.pages.FillBlankFormPage;
-import org.odk.collect.android.espressoutils.pages.FormEntryPage;
-import org.odk.collect.android.espressoutils.pages.GeneralSettingsPage;
-import org.odk.collect.android.espressoutils.pages.MainMenuPage;
+import org.odk.collect.android.support.pages.BlankFormSearchPage;
+import org.odk.collect.android.support.pages.ExitFormDialog;
+import org.odk.collect.android.support.pages.FillBlankFormPage;
+import org.odk.collect.android.support.pages.FormEntryPage;
+import org.odk.collect.android.support.pages.GeneralSettingsPage;
+import org.odk.collect.android.support.pages.MainMenuPage;
 import org.odk.collect.android.provider.InstanceProviderAPI;
 import org.odk.collect.android.support.ActivityHelpers;
 import org.odk.collect.android.support.CopyFormRule;
