+++ collect_app/src/androidTest/java/org/odk/collect/android/regression/ServerOtherTest.java
@@ -6,7 +6,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.RuleChain;
 import org.junit.runner.RunWith;
-import org.odk.collect.android.espressoutils.pages.MainMenuPage;
+import org.odk.collect.android.support.pages.MainMenuPage;
 import org.odk.collect.android.support.ResetStateRule;
 
 // Issue number NODK-235
