+++ collect_app/src/androidTest/java/org/odk/collect/android/support/pages/EditSavedFormPage.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.odk.collect.android.espressoutils.pages;
+package org.odk.collect.android.support.pages;
 
 import android.widget.RelativeLayout;
