+++ collect_app/src/androidTest/java/org/odk/collect/android/support/pages/ExitFormDialog.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.espressoutils.pages;
+package org.odk.collect.android.support.pages;
 
 import androidx.test.rule.ActivityTestRule;
